Use skills.ini instead of skills.tex
Maintaining skills.tex is difficult at best.  It is not readable, and it
is very hard to make changes.  The new skillUtil.py script converts the
ini format to the config format.  This allows changes to be made to the
skills.ini file and then converted automatically via the Makefile.